+++ lib/utils/log/handle_state.go
@@ -14,7 +14,6 @@ import (
 	"time"
 
 	"github.com/gravitational/trace"
-	"github.com/sirupsen/logrus"
 
 	"github.com/gravitational/teleport"
 )
@@ -108,13 +107,6 @@ func (s *handleState) appendAttr(a slog.Attr) bool {
 		nonEmpty := false
 		switch fields := a.Value.Any().(type) {
 		case map[string]any:
-			for k, v := range fields {
-				if s.appendAttr(slog.Any(k, v)) {
-					nonEmpty = true
-				}
-			}
-			return nonEmpty
-		case logrus.Fields:
 			for k, v := range fields {
 				if s.appendAttr(slog.Any(k, v)) {
 					nonEmpty = true
