+++ lib/utils/log/slog.go
@@ -24,7 +24,10 @@ import (
 	"log/slog"
 	"reflect"
 	"strings"
+	"unicode"
 
+	"github.com/gravitational/trace"
+	"github.com/sirupsen/logrus"
 	oteltrace "go.opentelemetry.io/otel/trace"
 )
 
@@ -34,7 +37,55 @@ const (
 
 	// TraceLevelText is the text representation of Trace verbosity.
 	TraceLevelText = "TRACE"
+
+	noColor = -1
+	red     = 31
+	yellow  = 33
+	blue    = 36
+	gray    = 37
+	// LevelField is the log field that stores the verbosity.
+	LevelField = "level"
+	// ComponentField is the log field that stores the calling component.
+	ComponentField = "component"
+	// CallerField is the log field that stores the calling file and line number.
+	CallerField = "caller"
+	// TimestampField is the field that stores the timestamp the log was emitted.
+	TimestampField = "timestamp"
+	messageField   = "message"
+	// defaultComponentPadding is a default padding for component field
+	defaultComponentPadding = 11
+	// defaultLevelPadding is a default padding for level field
+	defaultLevelPadding = 4
 )
+
+// SupportedLevelsText lists the supported log levels in their text
+// representation. All strings are in uppercase.
+var SupportedLevelsText = []string{
+	TraceLevelText,
+	slog.LevelDebug.String(),
+	slog.LevelInfo.String(),
+	slog.LevelWarn.String(),
+	slog.LevelError.String(),
+}
+
+// SlogLevelToLogrusLevel converts a [slog.Level] to its equivalent
+// [logrus.Level].
+func SlogLevelToLogrusLevel(level slog.Level) logrus.Level {
+	switch level {
+	case TraceLevel:
+		return logrus.TraceLevel
+	case slog.LevelDebug:
+		return logrus.DebugLevel
+	case slog.LevelInfo:
+		return logrus.InfoLevel
+	case slog.LevelWarn:
+		return logrus.WarnLevel
+	case slog.LevelError:
+		return logrus.ErrorLevel
+	default:
+		return logrus.FatalLevel
+	}
+}
 
 // DiscardHandler is a [slog.Handler] that discards all messages. It
 // is more efficient than a [slog.Handler] which outputs to [io.Discard] since
@@ -65,6 +116,45 @@ func addTracingContextToRecord(ctx context.Context, r *slog.Record) {
 
 	if spanContext.HasSpanID() {
 		r.AddAttrs(slog.String(spanID, spanContext.SpanID().String()))
+	}
+}
+
+var defaultFormatFields = []string{LevelField, ComponentField, CallerField, TimestampField}
+
+var knownFormatFields = map[string]struct{}{
+	LevelField:     {},
+	ComponentField: {},
+	CallerField:    {},
+	TimestampField: {},
+}
+
+func ValidateFields(formatInput []string) (result []string, err error) {
+	for _, component := range formatInput {
+		component = strings.TrimSpace(component)
+		if _, ok := knownFormatFields[component]; !ok {
+			return nil, trace.BadParameter("invalid log format key: %q", component)
+		}
+		result = append(result, component)
+	}
+	return result, nil
+}
+
+// needsQuoting returns true if any non-printable characters are found.
+func needsQuoting(text string) bool {
+	for _, r := range text {
+		if !unicode.IsPrint(r) {
+			return true
+		}
+	}
+	return false
+}
+
+func padMax(in string, chars int) string {
+	switch {
+	case len(in) < chars:
+		return in + strings.Repeat(" ", chars-len(in))
+	default:
+		return in[:chars]
 	}
 }
