+++ lib/utils/syslog_windows.go
@@ -22,13 +22,7 @@ import (
 	"io"
 
 	"github.com/gravitational/trace"
-	"github.com/sirupsen/logrus"
 )
-
-// NewSyslogHook always returns an error on Windows.
-func NewSyslogHook(io.Writer) (logrus.Hook, error) {
-	return nil, trace.NotImplemented("cannot use syslog on Windows")
-}
 
 // NewSyslogWriter always returns an error on Windows.
 func NewSyslogWriter() (io.Writer, error) {
